Refactor application settings into more minimal Config dataclass
Instead of using pydantic-settings, this commit transitions the
application into using homemade TOML parser due to its simplicity and
being more explicit what it does instead of having to know how
pydantic-settings actually works. The Config class also documents itself
now properly with docstrings and comments.

This commit introduces a client.py module which does the authentication
to a GitLab instance and instantiates the gitlab.Gitlab object for the
remainder of the glcli run.