Move towards drilldown method for querying through APIs
Needs a lot more work but at least the base principle works. In the
future, there needs to be some kind of caching for continous fetching of
the same objects.

Next, we need to implement a better handling of the single objects
fetching with handle_get handler. Now it only prints the attributes of
the object but we probably want to drill down deeper if the single
object's manager has other nested managers, such as ProjectIssueManager
and ProjectIssueLinkManager.