Separate runner and explorer logic into their own modules
In the future runner.run() only handles what is needed to initialize the
whole running prompting etc. The function could include something else
in the future but for now, it just passes the Gitlab client to the
explorer.

The explorer now has the whole prompting logic which is done as a
recursion. The explorer handles all the prompts until the proper non-browse
handler is found.